Remove lib/utils, retaining WhitespaceAnalysis
