--- lib/utils/git_walker/metrics/file_size.rb
@@ -1,7 +1,0 @@
-module GitWalker
-  module Metrics
-    def self.file_size(filepath, _repo)
-      File.size(filepath)
-    end
-  end
-end
