--- lib/utils/git_walker/metrics/no_of_authors.rb
@@ -1,11 +1,0 @@
-require 'set'
-
-module GitWalker
-  module Metrics
-    def self.no_of_authors(filepath, repo)
-      repo.gblob(filepath).log.each_with_object(Set.new) do |commit, authors|
-        authors.add(commit.author.email)
-      end.count
-    end
-  end
-end
